fix(desktop): cap a linked file's reported size to u32 for Specta (S3-T11)
Specta forbids exporting u64/bigint types to TypeScript. The size is
shown to the person, not used to decide anything, so a file above
4 GiB is shown at its cap rather than exactly (matching the same
convention S3-T10 used for table dimensions).

Bindings regenerated by a debug launch, not edited by hand.

+++ apps/desktop/src-tauri/src/commands/projects/open/types.rs
@@ -117,7 +117,7 @@ impl From<ReadError> for OpenFailure {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Type)]
 #[serde(tag = "kind", rename_all = "camelCase")]
 pub enum Availability {
-    Available { size: u64 },
+    Available { size: u32 },
     Missing,
     RootUnresolved,
     RootFolderMissing,
